extricated event loop
- event loop extricated into discrete class and membered into Network
